Refactor all buckets
