+++ cmd/functions.go
@@ -5,9 +5,9 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	_ "github.com/vshn/appcat/v4/pkg/comp-functions/functions/cloudscalebucket"
-	_ "github.com/vshn/appcat/v4/pkg/comp-functions/functions/exoscalebucket"
-	_ "github.com/vshn/appcat/v4/pkg/comp-functions/functions/miniobucket"
+	_ "github.com/vshn/appcat/v4/pkg/comp-functions/functions/buckets/cloudscalebucket"
+	_ "github.com/vshn/appcat/v4/pkg/comp-functions/functions/buckets/exoscalebucket"
+	_ "github.com/vshn/appcat/v4/pkg/comp-functions/functions/buckets/miniobucket"
 	_ "github.com/vshn/appcat/v4/pkg/comp-functions/functions/vshnkeycloak"
 	_ "github.com/vshn/appcat/v4/pkg/comp-functions/functions/vshnmariadb"
 	_ "github.com/vshn/appcat/v4/pkg/comp-functions/functions/vshnminio"
