+++ pkg/comp-functions/functions/buckets/cloudscalebucket/cloudscalebucket.go
@@ -1,27 +1,28 @@
-package exoscalebucket
+package cloudscalebucket
 
 import (
 	"context"
 	"fmt"
 
 	xpv1 "github.com/crossplane/crossplane-runtime/apis/common/v1"
-	"github.com/crossplane/crossplane-runtime/pkg/resource"
 	xfnproto "github.com/crossplane/function-sdk-go/proto/v1beta1"
 	appcatv1 "github.com/vshn/appcat/v4/apis/v1"
+	"github.com/vshn/appcat/v4/pkg/comp-functions/functions/buckets/util"
 	"github.com/vshn/appcat/v4/pkg/comp-functions/runtime"
-	exoscalev1 "github.com/vshn/provider-exoscale/apis/exoscale/v1"
+	cloudscalev1 "github.com/vshn/provider-cloudscale/apis/cloudscale/v1"
+	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 const (
-	bucketResName = "exoscale-bucket"
-	iamResName    = "exoscale-iam"
+	bucketResName = "cloudscale-bucket"
+	userResName   = "cloudscale-user"
 )
 
-// ProvisionExoscalebucket will create a bucket in cloudscale.
+// ProvisionCloudscalebucket will create a bucket in cloudscale.
 // This function will leverage provider-cloudscale to deploy proper users
 // alongside the bucket.
-func ProvisionExoscalebucket(_ context.Context, svc *runtime.ServiceRuntime) *xfnproto.Result {
+func ProvisionCloudscalebucket(_ context.Context, svc *runtime.ServiceRuntime) *xfnproto.Result {
 
 	bucket := &appcatv1.ObjectBucket{}
 
@@ -58,13 +59,16 @@ func ProvisionExoscalebucket(_ context.Context, svc *runtime.ServiceRuntime) *xf
 
 func addBucket(svc *runtime.ServiceRuntime, bucket *appcatv1.ObjectBucket, config string) error {
 
-	mb := &exoscalev1.Bucket{
+	mb := &cloudscalev1.Bucket{
 		ObjectMeta: metav1.ObjectMeta{},
-		Spec: exoscalev1.BucketSpec{
-			ForProvider: exoscalev1.BucketParameters{
-				BucketDeletionPolicy: exoscalev1.BucketDeletionPolicy(bucket.Spec.Parameters.BucketDeletionPolicy),
-				Zone:                 bucket.Spec.Parameters.Region,
+		Spec: cloudscalev1.BucketSpec{
+			ForProvider: cloudscalev1.BucketParameters{
+				BucketDeletionPolicy: cloudscalev1.BucketDeletionPolicy(bucket.Spec.Parameters.BucketDeletionPolicy),
+				Region:               bucket.Spec.Parameters.Region,
 				BucketName:           bucket.Spec.Parameters.BucketName,
+				CredentialsSecretRef: v1.SecretReference{
+					Namespace: svc.Config.Data["providerSecretNamespace"],
+				},
 			},
 			ResourceSpec: xpv1.ResourceSpec{
 				ProviderConfigReference: &xpv1.Reference{
@@ -74,18 +78,19 @@ func addBucket(svc *runtime.ServiceRuntime, bucket *appcatv1.ObjectBucket, confi
 		},
 	}
 
-	objName := getBucketObjectName(svc, bucket, bucketResName, mb.DeepCopy())
+	objName := util.GetBucketObjectName(svc, bucket, bucketResName, mb.DeepCopy())
 
 	mb.ObjectMeta.Name = objName
+	mb.Spec.ForProvider.CredentialsSecretRef.Name = objName
 
 	return svc.SetDesiredComposedResourceWithName(mb, bucketResName)
 }
 
 func addUser(svc *runtime.ServiceRuntime, bucket *appcatv1.ObjectBucket, config string) error {
 
-	user := &exoscalev1.IAMKey{
+	user := &cloudscalev1.ObjectsUser{
 		ObjectMeta: metav1.ObjectMeta{},
-		Spec: exoscalev1.IAMKeySpec{
+		Spec: cloudscalev1.ObjectsUserSpec{
 			ResourceSpec: xpv1.ResourceSpec{
 				ProviderConfigReference: &xpv1.Reference{
 					Name: config,
@@ -94,18 +99,18 @@ func addUser(svc *runtime.ServiceRuntime, bucket *appcatv1.ObjectBucket, config 
 					Namespace: svc.Config.Data["providerSecretNamespace"],
 				},
 			},
-			ForProvider: exoscalev1.IAMKeyParameters{
-				Zone: bucket.Spec.Parameters.Region,
+			ForProvider: cloudscalev1.ObjectsUserParameters{
+				DisplayName: fmt.Sprintf("%s.%s", bucket.Labels["crossplane.io/claim-namespace"], bucket.Labels["crossplane.io/claim-name"]),
 			},
 		},
 	}
 
-	objName := getBucketObjectName(svc, bucket, iamResName, user.DeepCopy())
+	objName := util.GetBucketObjectName(svc, bucket, userResName, user.DeepCopy())
 
 	user.ObjectMeta.Name = objName
 	user.Spec.WriteConnectionSecretToReference.Name = objName
 
-	cd, err := svc.GetObservedComposedResourceConnectionDetails(iamResName)
+	cd, err := svc.GetObservedComposedResourceConnectionDetails(userResName)
 	if err != nil && err != runtime.ErrNotFound {
 		return err
 	}
@@ -114,12 +119,12 @@ func addUser(svc *runtime.ServiceRuntime, bucket *appcatv1.ObjectBucket, config 
 		svc.SetConnectionDetail(v, k)
 	}
 
-	return svc.SetDesiredComposedResourceWithName(user, iamResName)
+	return svc.SetDesiredComposedResourceWithName(user, userResName)
 }
 
 func populateEndpointConnectionDetails(svc *runtime.ServiceRuntime) error {
 
-	bucket := &exoscalev1.Bucket{}
+	bucket := &cloudscalev1.Bucket{}
 
 	err := svc.GetObservedComposedResource(bucket, bucketResName)
 	if err != nil && err == runtime.ErrNotFound {
@@ -133,21 +138,4 @@ func populateEndpointConnectionDetails(svc *runtime.ServiceRuntime) error {
 
 	return nil
 
-}
-
-// Legacy buckets where created with wrong object names that break nested
-// services.
-// This logic returns the new naming scheme, if there's no existing bucket CR.
-// If there's already a CR it will simply return it's name.
-// The `obj` parameter should always be a deepCopy of the original, otherwise
-// the pointer in the calling function will have all the fields populated. Which
-// can lead to unexpected side effects.
-func getBucketObjectName(svc *runtime.ServiceRuntime, bucket *appcatv1.ObjectBucket, resName string, obj resource.Managed) string {
-
-	err := svc.GetObservedComposedResource(obj, resName)
-	if err != nil {
-		return bucket.Spec.Parameters.BucketName
-	}
-
-	return obj.GetName()
 }
