+++ pkg/comp-functions/functions/buckets/cloudscalebucket/register.go
