+++ pkg/comp-functions/functions/buckets/exoscalebucket/register.go
