+++ pkg/comp-functions/functions/buckets/miniobucket/register.go
